Allow continue on errors (#102)
This PR contains a retry mechanism in case of a failure.
RGHibernate tries to write a whole batch as a single transaction.
When a failure occurs, it tries to execute the updates one by one. If an update fails, it is moved to a DLQ stream together with details about the error, where it can be reviewed manually.