Get password using rgtrigger (#106)
The PR allows to get the connector password using another RG.TRIGGER registration that was register separately. This allows the user to add a costume code for password retrieval. RGHIBERNATE pass the connector name to the trigger so the password retrieval could get different passwords to different connectors.
In case of failure, the default password that was set on the xml configuration is kept.